move `merge_descriptors` from `nu-command` to `nu-cmd-base`
