+++ crates/nu-cmd-base/src/formats/to/delimited.rs
@@ -1,0 +1,20 @@
+use indexmap::{indexset, IndexSet};
+use nu_protocol::Value;
+
+pub fn merge_descriptors(values: &[Value]) -> Vec<String> {
+    let mut ret: Vec<String> = vec![];
+    let mut seen: IndexSet<String> = indexset! {};
+    for value in values {
+        let data_descriptors = match value {
+            Value::Record { cols, .. } => cols.to_owned(),
+            _ => vec!["".to_string()],
+        };
+        for desc in data_descriptors {
+            if !desc.is_empty() && !seen.contains(&desc) {
+                seen.insert(desc.to_string());
+                ret.push(desc.to_string());
+            }
+        }
+    }
+    ret
+}
