+++ crates/nu-cmd-base/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod formats;
 pub mod input_handler;
 pub mod util;
