+++ crates/nu-command/Cargo.toml
@@ -77,7 +77,6 @@ rayon = "1.7"
 regex = "1.7"
 roxmltree = "0.18"
 rusqlite = { version = "0.29", features = ["bundled"], optional = true }
-rust-embed = "6.7"
 same-file = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
